+++ squadcastv1/src/funcs/workflowsUpdateAction.ts
@@ -21,7 +21,6 @@ import * as errors from "../models/errors/index.js";
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import { SquadcastSDKError } from "../models/errors/squadcastsdkerror.js";
-import * as models from "../models/index.js";
 import * as operations from "../models/operations/index.js";
 import { APICall, APIPromise } from "../types/async.js";
 import { Result } from "../types/fp.js";
@@ -38,7 +37,7 @@ export function workflowsUpdateAction(
   options?: RequestOptions,
 ): APIPromise<
   Result<
-    models.V3WorkflowsActionResponse,
+    operations.WorkflowsUpdateWorkflowActionResponse,
     | errors.BadRequestError
     | errors.UnauthorizedError
     | errors.PaymentRequiredError
@@ -74,7 +73,7 @@ async function $do(
 ): Promise<
   [
     Result<
-      models.V3WorkflowsActionResponse,
+      operations.WorkflowsUpdateWorkflowActionResponse,
       | errors.BadRequestError
       | errors.UnauthorizedError
       | errors.PaymentRequiredError
@@ -198,7 +197,7 @@ async function $do(
   };
 
   const [result] = await M.match<
-    models.V3WorkflowsActionResponse,
+    operations.WorkflowsUpdateWorkflowActionResponse,
     | errors.BadRequestError
     | errors.UnauthorizedError
     | errors.PaymentRequiredError
@@ -219,7 +218,7 @@ async function $do(
     | UnexpectedClientError
     | SDKValidationError
   >(
-    M.json(200, models.V3WorkflowsActionResponse$inboundSchema),
+    M.json(200, operations.WorkflowsUpdateWorkflowActionResponse$inboundSchema),
     M.jsonErr(400, errors.BadRequestError$inboundSchema),
     M.jsonErr(401, errors.UnauthorizedError$inboundSchema),
     M.jsonErr(402, errors.PaymentRequiredError$inboundSchema),